// frontend/src/types/api.ts
export interface NavItem {
  name: string;
  path: string;
  icon: string;
}